finish out ReelGaps unit test
